requirements.txt
discord.py
python-dotenv
logging
asyncio
datetime